Refer: props, styles 파일 구분
