+++ packages/ui/src/buttons/button/button.props.ts
@@ -1,0 +1,10 @@
+export interface ButtonProps {
+  label: string;
+  disabled?: boolean;
+  onClick?: () => void;
+}
+
+export interface ButtonContainerProps{
+  isLoading?: boolean;
+  isDisabled?: boolean;
+}
