+++ packages/ui/src/buttons/index.ts
@@ -1,1 +1,1 @@
-export * from "./button";
+export * from "./button/button";
